Module zcash_client_backend::data_api

Interfaces for wallet data persistence & low-level wallet utilities.

Modules

chain

Tools for blockchain validation & scanning

error

Types for wallet error handling.

testing

wallet

Structs

DecryptedTransaction

A transaction that was detected during scanning of the blockchain, including its decrypted Sapling outputs.

PrunedBlock

The subset of information that is relevant to this wallet that has been decrypted and extracted from a CompactBlock.

SentTransaction

A transaction that was constructed and sent by the wallet.

SentTransactionOutput

Traits

BlockSource

This trait provides sequential access to raw blockchain data via a callback-oriented API.

WalletRead

Read-only operations required for light wallet functions.

WalletReadTransparent

WalletWrite

This trait encapsulates the write capabilities required to update stored wallet data.

WalletWriteTransparent