Module zcash_client_backend::data_api::chain

Tools for blockchain validation & scanning

Examples

use zcash_primitives::{
    consensus::{BlockHeight, Network, Parameters}
};

use zcash_client_backend::{
    data_api::{
        WalletRead, WalletWrite,
        chain::{
            BlockSource,
            error::Error,
            scan_cached_blocks,
            validate_chain,
            testing as chain_testing,
        },
        testing,
    },
};

let network = Network::TestNetwork;
let block_source = chain_testing::MockBlockSource;
let mut db_data = testing::MockWalletDb {
    network: Network::TestNetwork
};

// 1) Download new CompactBlocks into block_source.

// 2) Run the chain validator on the received blocks.
//
// Given that we assume the server always gives us correct-at-the-time blocks, any
// errors are in the blocks we have previously cached or scanned.
let max_height_hash = db_data.get_max_height_hash().map_err(Error::Wallet)?;
if let Err(e) = validate_chain(&network, &block_source, max_height_hash) {
    match e {
        Error::Chain(e) => {
            // a) Pick a height to rewind to.
            //
            // This might be informed by some external chain reorg information, or
            // heuristics such as the platform, available bandwidth, size of recent
            // CompactBlocks, etc.
            let rewind_height = e.at_height() - 10;

            // b) Rewind scanned block information.
            db_data.rewind_to_height(rewind_height);

            // c) Delete cached blocks from rewind_height onwards.
            //
            // This does imply that assumed-valid blocks will be re-downloaded, but it
            // is also possible that in the intervening time, a chain reorg has
            // occurred that orphaned some of those blocks.

            // d) If there is some separate thread or service downloading
            // CompactBlocks, tell it to go back and download from rewind_height
            // onwards.
        },
        e => {
            // handle or return other errors

        }
    }
}

// 3) Scan (any remaining) cached blocks.
//
// At this point, the cache and scanned data are locally consistent (though not
// necessarily consistent with the latest chain tip - this would be discovered the
// next time this codepath is executed after new blocks are received).
scan_cached_blocks(&network, &block_source, &mut db_data, None)

Modules

error

Types for chain scanning error handling.

testing

Traits

BlockSource

This trait provides sequential access to raw blockchain data via a callback-oriented API.

Functions

scan_cached_blocks

Scans at most limit new blocks added to the block source for any transactions received by the tracked accounts.

validate_chain

Checks that the scanned blocks in the data database, when combined with the recent CompactBlocks in the block_source database, form a valid chain.