Function zcash_client_backend::data_api::chain::scan_cached_blocks

pub fn scan_cached_blocks<ParamsT, DbT, BlockSourceT>(
    params: &ParamsT,
    block_source: &BlockSourceT,
    data_db: &mut DbT,
    limit: Option<u32>
) -> Result<(), Error<DbT::Error, BlockSourceT::Error, DbT::NoteRef>>where
    ParamsT: Parameters + Send + 'static,
    BlockSourceT: BlockSource,
    DbT: WalletWrite,

Scans at most limit new blocks added to the block source for any transactions received by the tracked accounts.

This function will return without error after scanning at most limit new blocks, to enable the caller to update their UI with scanning progress. Repeatedly calling this function will process sequential ranges of blocks, and is equivalent to calling scan_cached_blocks and passing None for the optional limit value.

This function pays attention only to cached blocks with heights greater than the highest scanned block in data. Cached blocks with lower heights are not verified against previously-scanned blocks. In particular, this function assumes that the caller is handling rollbacks.

For brand-new light client databases, this function starts scanning from the Sapling activation height. This height can be fast-forwarded to a more recent block by initializing the client database with a starting block (for example, calling init_blocks_table before this function if using zcash_client_sqlite).

Scanned blocks are required to be height-sequential. If a block is missing from the block source, an error will be returned with cause error::Cause::BlockHeightDiscontinuity.