Function zcash_client_backend::data_api::chain::validate_chain

pub fn validate_chain<ParamsT, BlockSourceT>(
    parameters: &ParamsT,
    block_source: &BlockSourceT,
    validate_from: Option<(BlockHeight, BlockHash)>
) -> Result<(), Error<Infallible, BlockSourceT::Error, Infallible>>where
    ParamsT: Parameters,
    BlockSourceT: BlockSource,

Checks that the scanned blocks in the data database, when combined with the recent CompactBlocks in the block_source database, form a valid chain.

This function is built on the core assumption that the information provided in the block source is more likely to be accurate than the previously-scanned information. This follows from the design (and trust) assumption that the lightwalletd server provides accurate block information as of the time it was requested.

Arguments:

parameters Network parameters

block_source Source of compact blocks

from_tip Height & hash of last validated block; if no validation has previously been performed, this will begin scanning from sapling_activation_height - 1

Returns:

Ok(()) if the combined chain is valid.

Err(Error::Chain(cause)) if the combined chain is invalid.

Err(e) if there was an error during validation unrelated to chain validity.

This function does not mutate either of the databases.