zcash_client_backend::data_api::error

Enum Error

pub enum Error<DataSourceError, CommitmentTreeError, SelectionError, FeeError, ChangeErrT, NoteRefT> {

    DataSource(DataSourceError),
    CommitmentTree(ShardTreeError<CommitmentTreeError>),
    NoteSelection(SelectionError),
    Change(ChangeError<ChangeErrT, NoteRefT>),
    Proposal(ProposalError),
    ProposalNotSupported,
    AccountIdNotRecognized,
    KeyNotRecognized,
    AccountCannotSpend,
    BalanceError(BalanceError),
    InsufficientFunds {
        available: Zatoshis,
        required: Zatoshis,
    },
    ScanRequired,
    Builder(Error<FeeError>),
    MemoForbidden,
    UnsupportedChangeType(PoolType),
    NoSupportedReceivers(Box<UnifiedAddress>),
    KeyNotAvailable(PoolType),
    NoteMismatch(NoteId),
    Address(ConversionError<&'static str>),
    AddressNotRecognized(TransparentAddress),
    PaysEphemeralTransparentAddress(String),

}

Errors that can occur as a consequence of wallet operations.

Variants

DataSource(DataSourceError)

An error occurred retrieving data from the underlying data source

CommitmentTree(ShardTreeError<CommitmentTreeError>)

An error in computations involving the note commitment trees.

NoteSelection(SelectionError)

An error in note selection

Change(ChangeError<ChangeErrT, NoteRefT>)

An error in change selection during transaction proposal construction

Proposal(ProposalError)

An error in transaction proposal construction

ProposalNotSupported

The proposal was structurally valid, but tried to do one of these unsupported things:

spend a prior shielded output;

pay to an output pool for which the corresponding feature is not enabled;

pay to a TEX address if the โtransparent-inputsโ feature is not enabled.

AccountIdNotRecognized

No account could be found corresponding to a provided ID.

KeyNotRecognized

No account could be found corresponding to a provided spending key.

AccountCannotSpend

The given account cannot be used for spending, because it is unable to maintain an accurate balance.

BalanceError(BalanceError)

Zcash amount computation encountered an overflow or underflow.

InsufficientFunds

Unable to create a new spend because the wallet balance is not sufficient.

Fields

available: Zatoshis

required: Zatoshis

ScanRequired

The wallet must first perform a scan of the blockchain before other operations can be performed.

Builder(Error<FeeError>)

An error occurred building a new transaction.

MemoForbidden

It is forbidden to provide a memo when constructing a transparent output.

UnsupportedChangeType(PoolType)

Attempted to send change to an unsupported pool.

This is indicative of a programming error; execution of a transaction proposal that presumes support for the specified pool was performed using an application that does not provide such support.

NoSupportedReceivers(Box<UnifiedAddress>)

Attempted to create a spend to an unsupported Unified Address receiver

KeyNotAvailable(PoolType)

A proposed transaction cannot be built because it requires spending an input of a type for which a key required to construct the transaction is not available.

NoteMismatch(NoteId)

A note being spent does not correspond to either the internal or external full viewing key for an account.

Address(ConversionError<&'static str>)

An error occurred parsing the address from a payment request.

AddressNotRecognized(TransparentAddress)

Available on crate feature transparent-inputs only.

The address associated with a record being inserted was not recognized as belonging to the wallet.

PaysEphemeralTransparentAddress(String)

Available on crate feature transparent-inputs only.

The wallet tried to pay to an ephemeral transparent address as a normal output.