zcash_client_backend::data_api

Module ll

Low-level data API

This module provides default implementations for several common client operations that rely on lower-level and more granular access to data. Client implementers should consider using the utility functions in this module when implementing traits such as WalletRead and WalletWrite in order to provide a consistent and robust user experience.

Modules

wallet

Traits

LowLevelWalletRead

A capability trait that provides low-level wallet database read operations. These operations are used to provide standard implementations for certain WalletWrite trait methods.

LowLevelWalletWrite

A capability trait that provides low-level wallet write operations. These operations are used to provide standard implementations for certain WalletWrite trait methods.

ReceivedOrchardOutputorchard

This trait provides a generalization over shielded Orchard output representations.

ReceivedSaplingOutput

This trait provides a generalization over shielded Sapling output representations.

TxMeta

A trait for types that can provide information about outputs spent by and fees that were paid for a given transaction.