zcash_client_backend::data_api::ll

Trait ReceivedOrchardOutput

pub trait ReceivedOrchardOutput {
    type AccountId;

    // Required methods
    fn index(&self) -> usize;
    fn account_id(&self) -> Self::AccountId;
    fn note(&self) -> &Note;
    fn memo(&self) -> Option<&MemoBytes>;
    fn is_change(&self) -> bool;
    fn nullifier(&self) -> Option<&Nullifier>;
    fn note_commitment_tree_position(&self) -> Option<Position>;
    fn recipient_key_scope(&self) -> Option<Scope>;
}

Available on crate feature orchard only.

This trait provides a generalization over shielded Orchard output representations.

Required Associated Types

type AccountId

Required Methods

fn index(&self) -> usize

Returns the index of the Orchard action that produced this output within the Orchard bundle.

fn account_id(&self) -> Self::AccountId

Returns the account ID for the account that received this output.

fn note(&self) -> &Note

Returns the received note.

fn memo(&self) -> Option<&MemoBytes>

Returns any memo associated with the output.

fn is_change(&self) -> bool

Returns whether or not the received output is counted as wallet-internal change, for the purpose of display.

fn nullifier(&self) -> Option<&Nullifier>

Returns the nullifier that will be revealed when the note is spent, if the output was observed using a key that provides the capability for nullifier computation.

fn note_commitment_tree_position(&self) -> Option<Position>

Returns the position of the note in the note commitment tree, if the transaction that produced the output has been mined.

fn recipient_key_scope(&self) -> Option<Scope>

Returns the HD derivation scope of the viewing key that decrypted the note, if known.