zcash_client_backend::data_api::ll::wallet

Function put_blocks

pub fn put_blocks<DbT, SE, TE>(
    wallet_db: &mut DbT,
    from_state: &ChainState,
    blocks: Vec<ScannedBlock<DbT::AccountId>>,
) -> Result<(), PutBlocksError<SE, TE>>
where
    DbT: LowLevelWalletWrite<Error = SE> + WalletCommitmentTrees<Error = TE>,
    DbT::TxRef: Eq + Hash,

Adds information about a sequence of scanned blocks to the provided data store.

Parameters

wallet_db: A handle to the underlying data store.

from_state: The note commitment tree state as of the end of the last block prior to the first block in the provided block vector.

blocks: The scanned block data to be added to the data store. This vector must contain data for blocks in sequentially increasing height order; PutBlocksError::NonSequentialBlocks will be returned if this invariant is violated.