zcash_client_backend::data_api::ll::wallet

Trait PutBlocksDbT

pub trait PutBlocksDbT<SE, TE, AR>:
    LowLevelWalletWrite<Error = SE>
    + WalletCommitmentTrees<Error = TE>
    + AddressStore<Error = SE, AccountRef = AR> { }

Available on crate feature transparent-inputs only.

A trait alias capturing the database capabilities required by put_blocks.

When the transparent-inputs feature is enabled, this additionally requires AddressStore so that transparent gap addresses can be maintained as new blocks are scanned.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.