zcash_client_backend::data_api::ll::wallet

Trait StoreDecryptedTxDbT

pub trait StoreDecryptedTxDbT: LowLevelWalletWrite + AddressStore<Error = <Self as LowLevelWalletRead>::Error, AccountRef = <Self as LowLevelWalletRead>::AccountRef>
where
    <Self as LowLevelWalletRead>::Error: From<GapAddressesError<<Self as LowLevelWalletRead>::Error>>,
{ }

Available on crate feature transparent-inputs only.

A trait alias capturing the database capabilities required by store_decrypted_tx.

When the transparent-inputs feature is enabled, this additionally requires AddressStore so that transparent gap addresses can be regenerated after storing a decrypted transaction.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.