zcash_client_backend::data_api::testing

Module pool

Available on crate features test-dependencies only.

Traits

ShieldedPoolTester

Trait that exposes the pool-specific types and operations necessary to run the single-shielded-pool tests on a given pool.

Functions

birthday_in_anchor_shard

change_note_spends_succeed

checkpoint_gaps

create_to_address_fails_on_incorrect_usk

data_db_truncation

external_address_change_spends_detected_in_restore_from_seed

fails_to_send_max_spendable_to_transparent_with_memotransparent-inputs

Tests that sending all the spendable funds within the given shielded pool in a single transaction to a transparent address with a memo fails.

fully_funded_fully_privateorchard

fully_funded_send_to_torchard and transparent-inputs

invalid_chain_cache_disconnected

metadata_queries_exclude_unwanted_notes

multi_pool_checkpointorchard

multi_pool_checkpoints_with_pruningorchard

ovk_policy_prevents_recovery_from_chain

pool_crossing_requiredorchard

proposal_fails_if_not_all_ephemeral_outputs_consumedtransparent-inputs

proposal_fails_with_no_blocks

receive_two_notes_with_same_value

Tests that the wallet correctly reports balance with two notes that are identical other than their note randomness.

reorg_to_checkpoint

scan_cached_blocks_allows_blocks_out_of_order

scan_cached_blocks_detects_spends_out_of_order

scan_cached_blocks_finds_change_notes

scan_cached_blocks_finds_received_notes

send_max_spendable_proposal_succeeds_when_unconfirmed_funds_present

Tests that attempting to send MaxSpendable funds within the given shielded pool in a single transaction succeeds if there are funds that are not yet confirmed.

send_multi_step_proposed_transfertransparent-inputs

send_single_step_proposed_transfer

Tests sending funds within the given shielded pool in a single transaction.

send_with_multiple_change_outputs

shield_transparenttransparent-inputs

spend_all_funds_multi_step_proposed_transfertransparent-inputs

Test attempts to sweep a wallet into a TEX address

spend_all_funds_single_step_proposed_transfer

Tests spending all funds within the given shielded pool in a single transaction.

spend_everything_multi_step_many_notes_proposed_transfertransparent-inputs

This test attempts to send the max spendable funds to a TEX address recipient checks that the transactions were stored and that the amounts involved are correct

spend_everything_multi_step_single_note_proposed_transfertransparent-inputs

This test attempts to send the max spendable funds to a TEX address recipient checks that the transactions were stored and that the amounts involved are correct

spend_everything_multi_step_with_marginal_notes_proposed_transfertransparent-inputs

This test attempts to send the max spendable funds to a TEX address recipient. The wallet contains many notes denominated with the marginal fee value. Checks that the transactions were stored and that the amounts involved are correct

spend_everything_proposal_fails_when_unconfirmed_funds_present

Tests that attempting to send all the spendable funds within the given shielded pool in a single transaction fail if there are funds that are not yet confirmed.

spend_everything_single_step_proposed_transfer

Tests sending every piece of spendable funds within the given shielded pool in a single transaction.

spend_fails_on_locked_notes

spend_fails_on_unverified_notes

spend_max_spendable_single_step_proposed_transfer

Tests max spendable funds within the given shielded pool in a single transaction.

spend_succeeds_to_t_addr_zero_change

valid_chain_states

wallet_recovery_computes_feestransparent-inputs

Ensure that wallet recovery recomputes fees.

zip317_spend

zip_315_confirmations_test_steps

Tests that inputs from a source can be spent according to the default ConfirmationsPolicy.