zcash_client_backend::data_api::testing::pool

Function spend_everything_multi_step_with_marginal_notes_proposed_transfer

pub fn spend_everything_multi_step_with_marginal_notes_proposed_transfer<T: ShieldedPoolTester, Dsf>(
    ds_factory: Dsf,
    cache: impl TestCache,
)
where
    Dsf: DataStoreFactory,
    <Dsf as DataStoreFactory>::AccountId: Debug,

Available on (crate features test-dependencies) and crate feature transparent-inputs only.

This test attempts to send the max spendable funds to a TEX address recipient. The wallet contains many notes denominated with the marginal fee value. Checks that the transactions were stored and that the amounts involved are correct