zcash_client_backend::data_api::testing::pool

Function spend_everything_single_step_proposed_transfer

pub fn spend_everything_single_step_proposed_transfer<T: ShieldedPoolTester>(
    dsf: impl DataStoreFactory,
    cache: impl TestCache,
)

Available on crate features test-dependencies only.

Tests sending every piece of spendable funds within the given shielded pool in a single transaction.

The test:

Adds funds to the wallet in a single note.

Checks that the wallet balances are correct.

Constructs a request to spend the whole balance to an external address in the same pool.

Builds the transaction.

Checks that the transaction was stored, and that the outputs are decryptable and have the expected details.