Trait zcash_client_backend::data_api::BlockSource

pub trait BlockSource {
    type Error;
    fn with_blocks<F>(
        &self, 
        from_height: BlockHeight, 
        limit: Option<u32>, 
        with_row: F
    ) -> Result<(), Self::Error>
    where
        F: FnMut(CompactBlock) -> Result<(), Self::Error>;
}

This trait provides sequential access to raw blockchain data via a callback-oriented API.

Associated Types

type Error

Required methods

fn with_blocks<F>(
&self,
from_height: BlockHeight,
limit: Option<u32>,
with_row: F
) -> Result<(), Self::Error> where
F: FnMut(CompactBlock) -> Result<(), Self::Error>,

Scan the specified limit number of blocks from the blockchain, starting at from_height, applying the provided callback to each block.