zcash_client_backend::data_api

Trait InputSource

pub trait InputSource {
    type Error: Debug;
    type AccountId: Copy + Debug + Eq + Hash;
    type NoteRef: Copy + Debug + Eq + Ord;

    // Required methods
    fn get_spendable_note(
        &self,
        txid: &TxId,
        protocol: ShieldedProtocol,
        index: u32,
        target_height: TargetHeight,
    ) -> Result<Option<ReceivedNote<Self::NoteRef, Note>>, Self::Error>;
    fn select_spendable_notes(
        &self,
        account: Self::AccountId,
        target_value: TargetValue,
        sources: &[ShieldedProtocol],
        target_height: TargetHeight,
        confirmations_policy: ConfirmationsPolicy,
        exclude: &[Self::NoteRef],
    ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error>;
    fn select_unspent_notes(
        &self,
        account: Self::AccountId,
        sources: &[ShieldedProtocol],
        target_height: TargetHeight,
        exclude: &[Self::NoteRef],
    ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error>;
    fn get_account_metadata(
        &self,
        account: Self::AccountId,
        selector: &NoteFilter,
        target_height: TargetHeight,
        exclude: &[Self::NoteRef],
    ) -> Result<AccountMeta, Self::Error>;

    // Provided methods
    fn get_unspent_transparent_output(
        &self,
        _outpoint: &OutPoint,
        _target_height: TargetHeight,
    ) -> Result<Option<WalletUtxo>, Self::Error> { ... }
    fn get_spendable_transparent_outputs(
        &self,
        _address: &TransparentAddress,
        _target_height: TargetHeight,
        _confirmations_policy: ConfirmationsPolicy,
    ) -> Result<Vec<WalletUtxo>, Self::Error> { ... }
}

A trait representing the capability to query a data store for unspent transaction outputs belonging to a account.

Required Associated Types

type Error: Debug

The type of errors produced by a wallet backend.

type AccountId: Copy + Debug + Eq + Hash

Backend-specific account identifier.

An account identifier corresponds to at most a single unified spending key’s worth of spend authority, such that both received notes and change spendable by that spending authority will be interpreted as belonging to that account. This might be a database identifier type or a UUID.

type NoteRef: Copy + Debug + Eq + Ord

Backend-specific note identifier.

For example, this might be a database identifier type or a UUID.

Required Methods

fn get_spendable_note( &self, txid: &TxId, protocol: ShieldedProtocol, index: u32, target_height: TargetHeight, ) -> Result<Option<ReceivedNote<Self::NoteRef, Note>>, Self::Error>

Fetches a spendable note by indexing into a transaction’s shielded outputs for the specified shielded protocol.

Returns Ok(None) if the note is not known to belong to the wallet or if the note is not spendable as of the given height.

fn select_spendable_notes( &self, account: Self::AccountId, target_value: TargetValue, sources: &[ShieldedProtocol], target_height: TargetHeight, confirmations_policy: ConfirmationsPolicy, exclude: &[Self::NoteRef], ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error>

Returns a list of spendable notes sufficient to cover the specified target value, if possible. Only spendable notes corresponding to the specified shielded protocol will be included.

fn select_unspent_notes( &self, account: Self::AccountId, sources: &[ShieldedProtocol], target_height: TargetHeight, exclude: &[Self::NoteRef], ) -> Result<ReceivedNotes<Self::NoteRef>, Self::Error>

Returns the list of notes belonging to the wallet that are unspent as of the specified target height.

fn get_account_metadata( &self, account: Self::AccountId, selector: &NoteFilter, target_height: TargetHeight, exclude: &[Self::NoteRef], ) -> Result<AccountMeta, Self::Error>

Returns metadata describing the structure of the wallet for the specified account.

The returned metadata value must exclude:

notes that are not considered spendable as of the given target_height

unspent notes excluded by the provided selector;

unspent notes identified in the given exclude list.

Implementations of this method may limit the complexity of supported queries. Such limitations should be clearly documented for the implementing type.

Provided Methods

fn get_unspent_transparent_output( &self, _outpoint: &OutPoint, _target_height: TargetHeight, ) -> Result<Option<WalletUtxo>, Self::Error>

Available on crate feature transparent-inputs only.

Fetches the transparent output corresponding to the provided outpoint if it is considered spendable as of the provided target_height.

Returns Ok(None) if the UTXO is not known to belong to the wallet or would not be spendable in a transaction mined in the block at the target height.

fn get_spendable_transparent_outputs( &self, _address: &TransparentAddress, _target_height: TargetHeight, _confirmations_policy: ConfirmationsPolicy, ) -> Result<Vec<WalletUtxo>, Self::Error>

Available on crate feature transparent-inputs only.

Returns the list of unspent transparent outputs received by this wallet at address such that, at height target_height:

the transaction that produced the output had or will have at least the required number of confirmations according to the provided ConfirmationsPolicy; and

the output can potentially be spent in a transaction mined in a block at the given target_height.

Any output that is potentially spent by an unmined transaction in the mempool should be excluded unless the spending transaction will be expired at target_height.