zcash_client_backend::data_api
Trait NoteRetention

pub trait NoteRetention<NoteRef> {
    // Required method
    fn should_retain_sapling(&self, note: &ReceivedNote<NoteRef, Note>) -> bool;
}

A predicate that can be used to choose whether or not a particular note is retained in note selection.

Required Methods

fn should_retain_sapling(&self, note: &ReceivedNote<NoteRef, Note>) -> bool

Returns whether the specified Sapling note should be retained.