zcash_client_backend::data_api

Trait WalletCommitmentTrees

pub trait WalletCommitmentTrees {
    type Error: Debug;
    type SaplingShardStore<'a>: ShardStore<H = Node, CheckpointId = BlockHeight, Error = Self::Error>;

    // Required methods
    fn with_sapling_tree_mut<F, A, E>(&mut self, callback: F) -> Result<A, E>
       where for<'a> F: FnMut(&'a mut ShardTree<Self::SaplingShardStore<'a>, { sapling::NOTE_COMMITMENT_TREE_DEPTH }, SAPLING_SHARD_HEIGHT>) -> Result<A, E>,
             E: From<ShardTreeError<Self::Error>>;
    fn put_sapling_subtree_roots(
        &mut self,
        start_index: u64,
        roots: &[CommitmentTreeRoot<Node>],
    ) -> Result<(), ShardTreeError<Self::Error>>;
}

This trait describes a capability for manipulating wallet note commitment trees.

Required Associated Types

type Error: Debug

type SaplingShardStore<'a>: ShardStore<H = Node, CheckpointId = BlockHeight, Error = Self::Error>

The type of the backing [ShardStore] for the Sapling note commitment tree.

Required Methods

fn with_sapling_tree_mut<F, A, E>(&mut self, callback: F) -> Result<A, E>
where for<'a> F: FnMut(&'a mut ShardTree<Self::SaplingShardStore<'a>, { sapling::NOTE_COMMITMENT_TREE_DEPTH }, SAPLING_SHARD_HEIGHT>) -> Result<A, E>, E: From<ShardTreeError<Self::Error>>,

Evaluates the given callback function with a reference to the Sapling note commitment tree maintained by the wallet.

fn put_sapling_subtree_roots( &mut self, start_index: u64, roots: &[CommitmentTreeRoot<Node>], ) -> Result<(), ShardTreeError<Self::Error>>

Adds a sequence of Sapling note commitment tree subtree roots to the data store.

Each such value should be the Merkle root of a subtree of the Sapling note commitment tree containing 2^SAPLING_SHARD_HEIGHT note commitments.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.