zcash_client_backend::data_api::wallet

Function create_pczt_from_proposal

pub fn create_pczt_from_proposal<DbT, ParamsT, InputsErrT, FeeRuleT, ChangeErrT, N>(
    wallet_db: &mut DbT,
    params: &ParamsT,
    account_id: <DbT as WalletRead>::AccountId,
    ovk_policy: OvkPolicy,
    proposal: &Proposal<FeeRuleT, N>,
) -> Result<Pczt, CreateErrT<DbT, InputsErrT, FeeRuleT, ChangeErrT, N>>
where
    DbT: WalletWrite + WalletCommitmentTrees,
    ParamsT: Parameters + Clone,
    FeeRuleT: FeeRule,
    DbT::AccountId: Serialize,

Available on crate feature pczt only.

Constructs a transaction using the inputs supplied by the given proposal.

Only single-step proposals are currently supported.

Returns a partially-created Zcash transaction (PCZT) that is ready to be authorized. You can use the following roles for this:

pczt::roles::prover::Prover

pczt::roles::signer::Signer (if you have local access to the spend authorizing keys)

pczt::roles::combiner::Combiner (if you create proofs and apply signatures in parallel)

Once the PCZT fully authorized, call extract_and_store_transaction_from_pczt to finish transaction creation.