zcash_client_backend::data_api::wallet

Function extract_and_store_transaction_from_pczt

pub fn extract_and_store_transaction_from_pczt<DbT, N>(
    wallet_db: &mut DbT,
    pczt: Pczt,
    sapling_vk: Option<(&SpendVerifyingKey, &OutputVerifyingKey)>,
    orchard_vk: Option<&VerifyingKey>,
) -> Result<TxId, ExtractErrT<DbT, N>>
where
    DbT: WalletWrite + WalletCommitmentTrees,
    DbT::AccountId: DeserializeOwned,

Available on crate feature pczt only.

Finalizes the given PCZT, and persists the transaction to the wallet database.

The PCZT should have been created via create_pczt_from_proposal, which adds metadata necessary for the wallet backend.

Returns the transaction ID for the resulting transaction.

sapling_vk is optional to allow the caller to check whether a PCZT has Sapling with pczt::roles::prover::Prover::requires_sapling_proofs, and avoid downloading the Sapling parameters if they are not needed. If sapling_vk is None, and the PCZT has a Sapling bundle, this function will return an error.

orchard_vk is optional to allow the caller to control where the Orchard verifying key is generated or cached. If orchard_vk is None, and the PCZT has an Orchard bundle, an Orchard verifying key will be generated on the fly.