zcash_client_backend::data_api::wallet

Function propose_send_max_transfer

pub fn propose_send_max_transfer<DbT, ParamsT, FeeRuleT, CommitmentTreeErrT>(
    wallet_db: &mut DbT,
    params: &ParamsT,
    spend_from_account: <DbT as InputSource>::AccountId,
    spend_pools: &[ShieldedProtocol],
    fee_rule: &FeeRuleT,
    recipient: ZcashAddress,
    memo: Option<MemoBytes>,
    mode: MaxSpendMode,
    confirmations_policy: ConfirmationsPolicy,
) -> Result<Proposal<FeeRuleT, <DbT as InputSource>::NoteRef>, ProposeSendMaxErrT<DbT, CommitmentTreeErrT, FeeRuleT>>
where
    DbT: WalletRead + InputSource<Error = <DbT as WalletRead>::Error>,
    <DbT as InputSource>::NoteRef: Copy + Eq + Ord,
    ParamsT: Parameters + Clone,
    FeeRuleT: FeeRule + Clone,

Select transaction inputs, compute fees, and construct a proposal for a transaction or series of transactions that would spend all available funds from the given spend_pools that can then be authorized and made ready for submission to the network with create_proposed_transactions.