Function zcash_client_backend::data_api::wallet::shield_transparent_funds

pub fn shield_transparent_funds<DbT, ParamsT, InputsT>(
    wallet_db: &mut DbT,
    params: &ParamsT,
    prover: impl SaplingProver,
    input_selector: &InputsT,
    shielding_threshold: NonNegativeAmount,
    usk: &UnifiedSpendingKey,
    from_addrs: &[TransparentAddress],
    memo: &MemoBytes,
    min_confirmations: u32
) -> Result<DbT::TxRef, Error<DbT::Error, InputsT::Error, <InputsT::FeeRule as FeeRule>::Error, DbT::NoteRef>>where
    ParamsT: Parameters,
    DbT: WalletWrite,
    DbT::NoteRef: Copy + Eq + Ord,
    InputsT: InputSelector<DataSource = DbT>,

Constructs a transaction that consumes available transparent UTXOs belonging to the specified secret key, and sends them to the default address for the provided Sapling extended full viewing key.

This procedure will not attempt to shield transparent funds if the total amount being shielded is less than the default fee to send the transaction. Fees will be paid only from the transparent UTXOs being consumed.

Parameters:

wallet_db: A read/write reference to the wallet database

params: Consensus parameters

prover: The sapling::TxProver to use in constructing the shielded transaction.

input_selector: The InputSelector to for note selection and change and fee determination

usk: The unified spending key that will be used to detect and spend transparent UTXOs, and that will provide the shielded address to which funds will be sent. Funds will be shielded to the internal (change) address associated with the most preferred shielded receiver corresponding to this account, or if no shielded receiver can be used for this account, this function will return an error. This procedure will return an error if the USK does not correspond to an account known to the wallet.

from_addrs: The list of transparent addresses that will be used to filter transaparent UTXOs received by the wallet. Only UTXOs received at one of the provided addresses will be selected to be shielded.

memo: A memo to be included in the output to the (internal) recipient. This can be used to take notes about auto-shielding operations internal to the wallet that the wallet can use to improve how it represents those shielding transactions to the user.

min_confirmations: The minimum number of confirmations that a previously received UTXO must have in the blockchain in order to be considered for being spent.