Function zcash_client_backend::data_api::wallet::spend

pub fn spend<E, N, P, D, R>(
    wallet_db: &mut D, 
    params: &P, 
    prover: impl TxProver, 
    extsk: &ExtendedSpendingKey, 
    account: AccountId, 
    request: &TransactionRequest, 
    ovk_policy: OvkPolicy, 
    min_confirmations: u32
) -> Result<R, E> where
    E: From<Error<N>>,
    P: Parameters + Clone,
    R: Copy + Debug,
    D: WalletWrite<Error = E, TxRef = R>, 

Constructs a transaction that sends funds as specified by the request argument and stores it to the wallet’s “sent transactions” data store, and returns a unique identifier for the transaction; this identifier is used only for internal reference purposes and is not the same as the transaction’s txid, although after v4 transactions have been made invalid in a future network upgrade, the txid could potentially be used for this type (as it is non-malleable for v5+ transactions).

This procedure uses the wallet’s underlying note selection algorithm to choose inputs of sufficient value to satisfy the request, if possible.

Do not call this multiple times in parallel, or you will generate transactions that double-spend the same notes.

Transaction privacy

ovk_policy specifies the desired policy for which outgoing viewing key should be able to decrypt the outputs of this transaction. This is primarily relevant to wallet recovery from backup; in particular, OvkPolicy::Discard will prevent the recipient’s address, and the contents of memo, from ever being recovered from the block chain. (The total value sent can always be inferred by the sender from the spent notes and received change.)

Regardless of the specified policy, create_spend_to_address saves to, value, and memo in db_data. This can be deleted independently of ovk_policy.

For details on what transaction information is visible to the holder of a full or outgoing viewing key, refer to ZIP 310.

Parameters:

wallet_db: A read/write reference to the wallet database

params: Consensus parameters

prover: The TxProver to use in constructing the shielded transaction.

extsk: The extended spending key that controls the funds that will be spent in the resulting transaction.

account: The ZIP32 account identifier associated with the extended spending key that controls the funds to be used in creating this transaction. This procedure will return an error if this does not correctly correspond to extsk.

request: The ZIP-321 payment request specifying the recipients and amounts for the transaction.

ovk_policy: The policy to use for constructing outgoing viewing keys that can allow the sender to view the resulting notes on the blockchain.

min_confirmations: The minimum number of confirmations that a previously received note must have in the blockchain in order to be considered for being spent. A value of 10 confirmations is recommended.