zcash_client_backend::data_api::wallet::input_selection

Trait ShieldingSelector

pub trait ShieldingSelector {
    type Error;
    type InputSource: InputSource;

    // Required method
    fn propose_shielding<ParamsT, ChangeT>(
        &self,
        params: &ParamsT,
        wallet_db: &Self::InputSource,
        change_strategy: &ChangeT,
        shielding_threshold: Zatoshis,
        source_addrs: &[TransparentAddress],
        to_account: <Self::InputSource as InputSource>::AccountId,
        target_height: BlockHeight,
        min_confirmations: u32,
    ) -> Result<Proposal<<ChangeT as ChangeStrategy>::FeeRule, Infallible>, InputSelectorError<<Self::InputSource as InputSource>::Error, Self::Error, ChangeT::Error, Infallible>>
       where ParamsT: Parameters,
             ChangeT: ChangeStrategy<MetaSource = Self::InputSource>;
}

Available on crate feature transparent-inputs only.

A strategy for selecting transaction inputs and proposing transaction outputs for shielding-only transactions (transactions which spend transparent UTXOs and send all transaction outputs to the wallet’s shielded internal address(es)).

Required Associated Types

type Error

The type of errors that may be generated in input selection

type InputSource: InputSource

The type of data source that the input selector expects to access to obtain input transparent UTXOs. This associated type permits input selectors that may use specialized knowledge of the internals of a particular backing data store, if the generic API of InputSource does not provide sufficiently fine-grained operations for a particular backing store to optimally perform input selection.

Required Methods

fn propose_shielding<ParamsT, ChangeT>( &self, params: &ParamsT, wallet_db: &Self::InputSource, change_strategy: &ChangeT, shielding_threshold: Zatoshis, source_addrs: &[TransparentAddress], to_account: <Self::InputSource as InputSource>::AccountId, target_height: BlockHeight, min_confirmations: u32, ) -> Result<Proposal<<ChangeT as ChangeStrategy>::FeeRule, Infallible>, InputSelectorError<<Self::InputSource as InputSource>::Error, Self::Error, ChangeT::Error, Infallible>>
where ParamsT: Parameters, ChangeT: ChangeStrategy<MetaSource = Self::InputSource>,

Performs input selection and returns a proposal for the construction of a shielding transaction.

Implementations should return the maximum possible number of economically useful inputs required to supply at least the requested value, choosing only inputs received at the specified source addresses. If insufficient funds are available to satisfy the required outputs for the shielding request, this operation must fail and return InputSelectorError::InsufficientFunds.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.