zcash_client_backend::fees

Enum ChangeError

pub enum ChangeError<E, NoteRefT> {
    InsufficientFunds {
        available: Zatoshis,
        required: Zatoshis,
    },
    DustInputs {
        transparent: Vec<OutPoint>,
        sapling: Vec<NoteRefT>,
    },
    StrategyError(E),
    BundleError(&'static str),
}

Errors that can occur in computing suggested change and/or fees.

Variants

InsufficientFunds

Insufficient inputs were provided to change selection to fund the required outputs and fees.

Fields

available: Zatoshis

The total of the inputs provided to change selection

required: Zatoshis

The total amount of input value required to fund the requested outputs, including the required fees.

DustInputs

Some of the inputs provided to the transaction have value less than the marginal fee, and could not be determined to have any economic value in the context of this input selection.

This determination is potentially conservative in the sense that inputs with value less than or equal to the marginal fee might be excluded, even though in practice they would not cause the fee to increase. Inputs with value greater than the marginal fee will never be excluded.

The ordering of the inputs in each list is unspecified.

Fields

transparent: Vec<OutPoint>

The outpoints for transparent inputs that could not be determined to have economic value in the context of this input selection.

sapling: Vec<NoteRefT>

The identifiers for Sapling inputs that could not be determined to have economic value in the context of this input selection.

StrategyError(E)

An error occurred that was specific to the change selection strategy in use.

BundleError(&'static str)

The proposed bundle structure would violate bundle type construction rules.