zcash_client_backend::fees::sapling

Trait BundleView

pub trait BundleView<NoteRef> {
    type In: InputView<NoteRef>;
    type Out: OutputView;

    // Required methods
    fn bundle_type(&self) -> BundleType;
    fn inputs(&self) -> &[Self::In];
    fn outputs(&self) -> &[Self::Out];
}

A trait that provides a minimized view of Sapling bundle configuration suitable for use in fee and change calculation.

Required Associated Types

type In: InputView<NoteRef>

The type of inputs to the bundle.

type Out: OutputView

The type of inputs of the bundle.

Required Methods

fn bundle_type(&self) -> BundleType

Returns the type of the bundle

fn inputs(&self) -> &[Self::In]

Returns the inputs to the bundle.

fn outputs(&self) -> &[Self::Out]

Returns the outputs of the bundle.