zcash_client_backend::fees::sapling

Trait OutputView

pub trait OutputView {
    // Required method
    fn value(&self) -> Zatoshis;
}

A trait that provides a minimized view of a Sapling output suitable for use in fee and change calculation.

Required Methods

fn value(&self) -> Zatoshis

The value of the output being produced.