zcash_client_backend::fees

Module standard

Change strategies designed for use with a standard fee.

Type Aliases

MultiOutputChangeStrategy

A change strategy that proposes change as potentially multiple evenly-sized outputs having at least a threshold value. The output pool is chosen as the most current pool that avoids unnecessary pool-crossing (with a specified fallback when the transaction has no shielded inputs). Fee calculation is delegated to the provided fee rule.

SingleOutputChangeStrategy

A change strategy that proposes change as a single output. The output pool is chosen as the most current pool that avoids unnecessary pool-crossing (with a specified fallback when the transaction has no shielded inputs). Fee calculation is delegated to the provided fee rule.