Trait zcash_client_backend::fees::ChangeStrategy

pub trait ChangeStrategy {
    type FeeRule: FeeRule;
    type Error;

    fn fee_rule(&self) -> &Self::FeeRule;
    fn compute_balance<P: Parameters, NoteRefT: Clone>(
        &self,
        params: &P,
        target_height: BlockHeight,
        transparent_inputs: &[impl InputView],
        transparent_outputs: &[impl OutputView],
        sapling_inputs: &[impl InputView<NoteRefT>],
        sapling_outputs: &[impl OutputView],
        dust_output_policy: &DustOutputPolicy
    ) -> Result<TransactionBalance, ChangeError<Self::Error, NoteRefT>>;
}

A trait that represents the ability to compute the suggested change and fees that must be paid by a transaction having a specified set of inputs and outputs.

Required Associated Types

type FeeRule: FeeRule

type Error

Required Methods

fn fee_rule(&self) -> &Self::FeeRule

Returns the fee rule that this change strategy will respect when performing balance computations.

fn compute_balance<P: Parameters, NoteRefT: Clone>(
&self,
params: &P,
target_height: BlockHeight,
transparent_inputs: &[impl InputView],
transparent_outputs: &[impl OutputView],
sapling_inputs: &[impl InputView<NoteRefT>],
sapling_outputs: &[impl OutputView],
dust_output_policy: &DustOutputPolicy
) -> Result<TransactionBalance, ChangeError<Self::Error, NoteRefT>>

Computes the totals of inputs, suggested change amounts, and fees given the provided inputs and outputs being used to construct a transaction.

The fee computed as part of this operation should take into account the prospective change outputs recommended by this operation. If insufficient funds are available to supply the requested outputs and required fees, implementations should return ChangeError::InsufficientFunds.