zcash_client_backend::fees

Trait MetaSource

pub trait MetaSource {
    type Error;
    type AccountId;
    type NoteRef;
}

A trait that defines a set of types used in wallet metadata retrieval. Ordinarily, this will correspond to a type that implements InputSource, and a blanket implementation of this trait is provided for all types that implement InputSource.

If more capabilities are required of the backend than are exposed in the InputSource trait, the implementer of this trait should define their own trait that descends from InputSource and adds the required capabilities there, and then implement that trait for their desired database backend.

Required Associated Types

type Error

type AccountId

type NoteRef