zcash_client_backend::fees::zip317

Struct MultiOutputChangeStrategy

pub struct MultiOutputChangeStrategy<R, I> { /* private fields */ }

A change strategy that attempts to split the change value into some number of equal-sized notes as dictated by the included SplitPolicy value.

Implementations

impl<R, I> MultiOutputChangeStrategy<R, I>

pub fn new( fee_rule: R, change_memo: Option<MemoBytes>, fallback_change_pool: ShieldedProtocol, dust_output_policy: DustOutputPolicy, split_policy: SplitPolicy, ) -> Self

Constructs a new MultiOutputChangeStrategy with the specified ZIP 317 fee parameters, change memo, and change splitting policy.

This change strategy will fall back to creating a single change output if insufficient change value is available to create notes with at least the minimum value dictated by the split policy.

fallback_change_pool: the pool to which change will be sent if when more than one shielded pool is enabled via feature flags, and the transaction has no shielded inputs.

split_policy: A policy value describing how the change value should be returned as multiple notes.