zcash_client_backend

Function decrypt_transaction

pub fn decrypt_transaction<'a, P: Parameters, AccountId: Copy>(
    params: &P,
    mined_height: Option<BlockHeight>,
    chain_tip_height: Option<BlockHeight>,
    tx: &'a Transaction,
    ufvks: &HashMap<AccountId, UnifiedFullViewingKey>,
) -> DecryptedTransaction<'a, AccountId>

Scans a [Transaction] for any information that can be decrypted by the set of UnifiedFullViewingKeys.

Parameters

params: The network parameters corresponding to the network the transaction was created for.

mined_height: The height at which the transaction was mined, or None for unmined transactions.

chain_tip_height: The current chain tip height, if known. This parameter will be unused if mined_height.is_some().

tx: The transaction to decrypt.

ufvks: The UnifiedFullViewingKeys to use in trial decryption, keyed by the identifiers for the wallet accounts they correspond to.