zcash_client_backend

Module proposal

Types related to the construction and evaluation of transaction proposals.

Structs

Proposal

A proposal for a series of transactions to be created.

ShieldedInputs

The Sapling inputs to a proposed transaction.

Step

The inputs to be consumed and outputs to be produced in a proposed transaction.

StepOutput

A reference to the output of a step in a proposal.

Enums

ProposalError

Errors that can occur in construction of a Step.

StepOutputIndex

A reference to either a payment or change output within a step.