zcash_client_backend::proto

Module compact_formats

Structs

ChainMetadata

Information about the state of the chain as of a given block.

CompactBlock

A compact representation of the shielded data in a Zcash block.

CompactOrchardAction

A compact representation of an Orchard Action.

CompactSaplingOutput

A compact representation of a Sapling Output.

CompactSaplingSpend

A compact representation of a Sapling Spend.

CompactTx

A compact representation of the shielded data in a Zcash transaction.