zcash_client_backend

Module scanning

Tools for scanning a compact representation of the Zcash block chain.

Modules

testingtest-dependencies

Structs

Nullifiers

The set of nullifiers being tracked by a wallet.

ScanningKey

An incoming viewing key, paired with an optional nullifier key and key source metadata.

ScanningKeys

A set of keys to be used in scanning for decryptable transaction outputs.

Enums

ScanError

Errors that may occur in chain scanning

Traits

ScanningKeyOps

A key that can be used to perform trial decryption and nullifier computation for a CompactSaplingOutput or CompactOrchardAction.

Functions

scan_block

Scans a CompactBlock with a set of ScanningKeys.