zcash_client_backend::scanning

Function scan_block

pub fn scan_block<P, AccountId, IvkTag>(
    params: &P,
    block: CompactBlock,
    scanning_keys: &ScanningKeys<AccountId, IvkTag>,
    nullifiers: &Nullifiers<AccountId>,
    prior_block_metadata: Option<&BlockMetadata>,
) -> Result<ScannedBlock<AccountId>, ScanError>
where
    P: Parameters + Send + 'static,
    AccountId: Default + Eq + Hash + ConditionallySelectable + Send + 'static,
    IvkTag: Copy + Hash + Eq + Send + 'static,

Scans a CompactBlock with a set of ScanningKeys.

Returns a vector of WalletTxs decryptable by any of the given keys. If an output is decrypted by a full viewing key, the nullifiers of that output will also be computed.