zcash_client_backend
Module sync

Available on crate feature sync only.

Implementation of the synchronization flow described in the crate root.

This is currently a simple implementation that does not yet implement a few features:

Block batches are not downloaded in parallel with scanning.

Transactions are not enhanced once detected (that is, after an output is detected in a transaction, the full transaction is not downloaded and scanned).

There is no mechanism for notifying the caller of progress updates.

There is no mechanism for interrupting the synchronization flow, other than ending the process.

Enums

Error

Errors that can occur while syncing.

Functions

run

Scans the chain until the wallet is up-to-date.