zcash_client_backend::tor::http

Module cryptex

Available on crate feature tor only.

Cryptocurrency exchange rate APIs.

Modules

exchanges

Exchanges for which we know how to query data over Tor.

Structs

ExchangeData

Data queried from an Exchange.

Exchanges

A set of Exchanges that can be queried for ZEC data.

ExchangesBuilder

Builder type for Exchanges.

Traits

Exchange

An exchange that can be queried for ZEC data.

LocalExchange

An exchange that can be queried for ZEC data.