zcash_client_backend::tor::http::cryptex

Module exchanges

Available on crate feature tor only.

Exchanges for which we know how to query data over Tor.

Queries to these exchanges will be retried a single time on error.

Structs

Binance

Querier for the Binance exchange.

CoinEx

Querier for the CoinEx exchange.

Coinbase

Querier for the Coinbase exchange.

DigiFinex

Querier for the DigiFinex exchange.

Gemini

Querier for the Gemini exchange.

Kraken

Querier for the Kraken exchange.

KuCoin

Querier for the KuCoin exchange.

Mexc

Querier for the MEXC exchange.

Xt

Querier for the XT exchange.