zcash_client_backend

Module wallet

Structs representing transaction data scanned from the block chain by a wallet or light client.

Structs

NoteId

A unique identifier for a shielded transaction output

ReceivedNote

A note that was received by the wallet, along with contextual information about the output that generated the note and the key that is required to spend it.

TransparentAddressMetadatatransparent-inputs

Information about a transparent address controlled by the wallet.

WalletOutput

An output that was successfully decrypted in the process of wallet scanning.

WalletSpend

A reference to a spent note belonging to the wallet within a transaction.

WalletTransparentOutput

A transparent output controlled by the wallet.

WalletTx

The shielded subset of a Transaction’s data that is relevant to a particular wallet.

Enums

Exposuretransparent-inputs

Metadata describing whether and when a transparent address was exposed by the wallet.

GapMetadata

Metadata describing the gap limit position of a transparent address.

Note

An enumeration of supported shielded note types for use in ReceivedNote

OvkPolicy

Describes a policy for which outgoing viewing key should be able to decrypt transaction outputs.

Recipient

A type that represents the recipient of a transaction output:

TransparentAddressSourcetransparent-inputs

Source information for a transparent address.

Type Aliases

WalletOrchardOutputorchard

The output part of an Orchard Action that was decrypted in the process of scanning.

WalletOrchardSpendorchard

A type alias for Orchard WalletSpends.

WalletSaplingOutput

A subset of an OutputDescription relevant to wallets and light clients.

WalletSaplingSpend

A type alias for Sapling WalletSpends.