Function zcash_client_backend::welding_rig::scan_block

pub fn scan_block<P: Parameters, K: ScanningKey>(
    params: &P, 
    block: CompactBlock, 
    vks: &[(&AccountId, &K)], 
    nullifiers: &[(AccountId, Nullifier)], 
    tree: &mut CommitmentTree<Node>, 
    existing_witnesses: &mut [&mut IncrementalWitness<Node>]
) -> Vec<WalletTx<K::Nf>>

Scans a CompactBlock with a set of ScanningKeys.

Returns a vector of WalletTxs belonging to any of the given ScanningKeys. If scanning with a full viewing key, the nullifiers of the resulting WalletShieldedOutputs will also be computed.

The given CommitmentTree and existing IncrementalWitnesses are incremented appropriately.

The implementation of ScanningKey may either support or omit the computation of the nullifiers for received notes; the implementation for ExtendedFullViewingKey will derive the nullifiers for received notes and return them as part of the resulting WalletShieldedOutputs, whereas the implementation for SaplingIvk cannot do so and will return the unit value in those outputs instead.