Trait zcash_client_backend::welding_rig::ScanningKey

pub trait ScanningKey {
    type Scope: Clone + Eq + Hash + Send + 'static;
    type SaplingNk: Clone;
    type SaplingKeys: IntoIterator<Item = (Self::Scope, SaplingIvk, Self::SaplingNk)>;
    type Nf;

    fn to_sapling_keys(&self) -> Self::SaplingKeys;
    fn sapling_nf(
        key: &Self::SaplingNk,
        note: &Note,
        witness: &IncrementalWitness<Node>
    ) -> Self::Nf;
}

A key that can be used to perform trial decryption and nullifier computation for a Sapling CompactSaplingOutput

The purpose of this trait is to enable scan_block and related methods to be used with either incoming viewing keys or full viewing keys, with the data returned from trial decryption being dependent upon the type of key used. In the case that an incoming viewing key is used, only the note and payment address will be returned; in the case of a full viewing key, the nullifier for the note can also be obtained.

Required Associated Types

type Scope: Clone + Eq + Hash + Send + 'static

The type representing the scope of the scanning key.

type SaplingNk: Clone

The type of key that is used to decrypt Sapling outputs;

type SaplingKeys: IntoIterator<Item = (Self::Scope, SaplingIvk, Self::SaplingNk)>

type Nf

The type of nullifier extracted when a note is successfully obtained by trial decryption.

Required Methods

fn to_sapling_keys(&self) -> Self::SaplingKeys

Obtain the underlying Sapling incoming viewing key(s) for this scanning key.

fn sapling_nf(
key: &Self::SaplingNk,
note: &Note,
witness: &IncrementalWitness<Node>
) -> Self::Nf

Produces the nullifier for the specified note and witness, if possible.

IVK-based implementations of this trait cannot successfully derive nullifiers, in which case Self::Nf should be set to the unit type and this function is a no-op.