Crate zcash_client_backend

A crate for implementing Zcash light clients.

zcash_client_backend contains Rust structs and traits for creating shielded Zcash light clients.

Modules

address

Structs for handling supported address types.

data_api

Interfaces for wallet data persistence & low-level wallet utilities.

encoding

Encoding and decoding functions for Zcash key and address structs.

keys

Helper functions for managing light client key material.

proto

Generated code for handling light client protobuf structs.

wallet

Structs representing transaction data scanned from the block chain by a wallet or light client.

welding_rig

Tools for scanning a compact representation of the Zcash block chain.

zip321

Reference implementation of the ZIP-321 standard for payment requests.

Structs

DecryptedOutput

A decrypted shielded output.

Functions

decrypt_transaction

Scans a Transaction for any information that can be decrypted by the set of ExtendedFullViewingKeys.