zcash_client_backend::data_api::chain

Function scan_cached_blocks

pub fn scan_cached_blocks<ParamsT, DbT, BlockSourceT>(
    params: &ParamsT,
    block_source: &BlockSourceT,
    data_db: &mut DbT,
    from_height: BlockHeight,
    from_state: &ChainState,
    limit: usize,
) -> Result<ScanSummary, Error<DbT::Error, BlockSourceT::Error>>
where
    ParamsT: Parameters + Send + 'static,
    BlockSourceT: BlockSource,
    DbT: WalletWrite,
    <DbT as WalletRead>::AccountId: ConditionallySelectable + Default + Send + 'static,

Scans at most limit blocks from the provided block source in order to find transactions received by the accounts tracked in the provided wallet database.

This function will return after scanning at most limit new blocks, to enable the caller to update their UI with scanning progress.

Panics

This method will panic if from_height != from_state.block_height() + 1.