Function zcash_client_backend::data_api::chain::scan_cached_blocks

pub fn scan_cached_blocks<E, N, P, C, D>(
    params: &P, 
    cache: &C, 
    data: &mut D, 
    limit: Option<u32>
) -> Result<(), E> where
    P: Parameters,
    C: BlockSource<Error = E>,
    D: WalletWrite<Error = E, NoteRef = N>,
    N: Copy + Debug,
    E: From<Error<N>>, 

Scans at most limit new blocks added to the cache for any transactions received by the tracked accounts.

This function will return without error after scanning at most limit new blocks, to enable the caller to update their UI with scanning progress. Repeatedly calling this function will process sequential ranges of blocks, and is equivalent to calling scan_cached_blocks and passing None for the optional limit value.

This function pays attention only to cached blocks with heights greater than the highest scanned block in data. Cached blocks with lower heights are not verified against previously-scanned blocks. In particular, this function assumes that the caller is handling rollbacks.

For brand-new light client databases, this function starts scanning from the Sapling activation height. This height can be fast-forwarded to a more recent block by initializing the client database with a starting block (for example, calling init_blocks_table before this function if using zcash_client_sqlite).

Scanned blocks are required to be height-sequential. If a block is missing from the cache, an error will be returned with kind ChainInvalid::BlockHeightDiscontinuity.

Examples

use tempfile::NamedTempFile;
use zcash_primitives::consensus::{
    Network,
    Parameters,
};
use zcash_client_backend::{
    data_api::chain::scan_cached_blocks,
};
use zcash_client_sqlite::{
    BlockDb,
    WalletDb,
    error::SqliteClientError,
    wallet::init::init_wallet_db,
};

let cache_file = NamedTempFile::new().unwrap();
let cache = BlockDb::for_path(cache_file).unwrap();

let data_file = NamedTempFile::new().unwrap();
let db_read = WalletDb::for_path(data_file, Network::TestNetwork)?;
init_wallet_db(&db_read)?;

let mut data = db_read.get_update_ops()?;
scan_cached_blocks(&Network::TestNetwork, &cache, &mut data, None)?;