Function zcash_client_backend::data_api::chain::validate_chain

pub fn validate_chain<N, E, P, C>(
    parameters: &P, 
    cache: &C, 
    validate_from: Option<(BlockHeight, BlockHash)>
) -> Result<(), E> where
    E: From<Error<N>>,
    P: Parameters,
    C: BlockSource<Error = E>, 

Checks that the scanned blocks in the data database, when combined with the recent CompactBlocks in the cache database, form a valid chain.

This function is built on the core assumption that the information provided in the cache database is more likely to be accurate than the previously-scanned information. This follows from the design (and trust) assumption that the lightwalletd server provides accurate block information as of the time it was requested.

Arguments:

parameters Network parameters

cache Source of compact blocks

from_tip Height & hash of last validated block; if no validation has previously been performed, this will begin scanning from sapling_activation_height - 1

Returns:

Ok(()) if the combined chain is valid.

Err(ErrorKind::InvalidChain(upper_bound, cause)) if the combined chain is invalid. upper_bound is the height of the highest invalid block (on the assumption that the highest block in the cache database is correct).

Err(e) if there was an error during validation unrelated to chain validity.

This function does not mutate either of the databases.