zcash_client_backend::data_api::testing

Module pool

Available on crate feature test-dependencies only.

Traits

ShieldedPoolTester

Trait that exposes the pool-specific types and operations necessary to run the single-shielded-pool tests on a given pool.

Functions

birthday_in_anchor_shard

change_note_spends_succeed

checkpoint_gaps

create_to_address_fails_on_incorrect_usk

data_db_truncation

external_address_change_spends_detected_in_restore_from_seed

invalid_chain_cache_disconnected

metadata_queries_exclude_unwanted_notes

ovk_policy_prevents_recovery_from_chain

proposal_fails_if_not_all_ephemeral_outputs_consumedtransparent-inputs

proposal_fails_with_no_blocks

reorg_to_checkpoint

scan_cached_blocks_allows_blocks_out_of_order

scan_cached_blocks_detects_spends_out_of_order

scan_cached_blocks_finds_change_notes

scan_cached_blocks_finds_received_notes

send_multi_step_proposed_transfertransparent-inputs

send_single_step_proposed_transfer

Tests sending funds within the given shielded pool in a single transaction.

send_with_multiple_change_outputs

shield_transparenttransparent-inputs

spend_fails_on_locked_notes

spend_fails_on_unverified_notes

spend_succeeds_to_t_addr_zero_change

valid_chain_states

zip317_spend