Trait zcash_client_backend::data_api::WalletWrite

pub trait WalletWrite: WalletRead {
    fn advance_by_block(
        &mut self, 
        block: &PrunedBlock<'_>, 
        updated_witnesses: &[(Self::NoteRef, IncrementalWitness<Node>)]
    ) -> Result<Vec<(Self::NoteRef, IncrementalWitness<Node>)>, Self::Error>;
    fn store_decrypted_tx(
        &mut self, 
        received_tx: &DecryptedTransaction<'_>
    ) -> Result<Self::TxRef, Self::Error>;
    fn store_sent_tx(
        &mut self, 
        sent_tx: &SentTransaction<'_>
    ) -> Result<Self::TxRef, Self::Error>;
    fn rewind_to_height(
        &mut self, 
        block_height: BlockHeight
    ) -> Result<(), Self::Error>;
}

This trait encapsulates the write capabilities required to update stored wallet data.

Required methods

fn advance_by_block(
&mut self,
block: &PrunedBlock<'_>,
updated_witnesses: &[(Self::NoteRef, IncrementalWitness<Node>)]
) -> Result<Vec<(Self::NoteRef, IncrementalWitness<Node>)>, Self::Error>

Updates the state of the wallet database by persisting the provided block information, along with the updated witness data that was produced when scanning the block for transactions pertaining to this wallet.

fn store_decrypted_tx(
&mut self,
received_tx: &DecryptedTransaction<'_>
) -> Result<Self::TxRef, Self::Error>

Caches a decrypted transaction in the persistent wallet store.

fn store_sent_tx(
&mut self,
sent_tx: &SentTransaction<'_>
) -> Result<Self::TxRef, Self::Error>

fn rewind_to_height(
&mut self,
block_height: BlockHeight
) -> Result<(), Self::Error>

Rewinds the wallet database to the specified height.

This method assumes that the state of the underlying data store is consistent up to a particular block height. Since it is possible that a chain reorg might invalidate some stored state, this method must be implemented in order to allow users of this API to “reset” the data store to correctly represent chainstate as of a specified block height.

After calling this method, the block at the given height will be the most recent block and all other operations will treat this block as the chain tip for balance determination purposes.

There may be restrictions on how far it is possible to rewind.