Module zcash_client_backend::data_api::wallet

Functions

create_spend_to_address

Creates a transaction paying the specified address from the given account.

decrypt_and_store_transaction

Scans a Transaction for any information that can be decrypted by the accounts in the wallet, and saves it to the wallet.

shield_transparent_funds

Constructs a transaction that consumes available transparent UTXOs belonging to the specified secret key, and sends them to the default address for the provided Sapling extended full viewing key.

spend

Constructs a transaction that sends funds as specified by the request argument and stores it to the wallet’s “sent transactions” data store, and returns a unique identifier for the transaction; this identifier is used only for internal reference purposes and is not the same as the transaction’s txid, although after v4 transactions have been made invalid in a future network upgrade, the txid could potentially be used for this type (as it is non-malleable for v5+ transactions).