Function zcash_client_backend::data_api::wallet::create_spend_to_address

pub fn create_spend_to_address<E, N, P, D, R>(
    wallet_db: &mut D, 
    params: &P, 
    prover: impl TxProver, 
    account: AccountId, 
    extsk: &ExtendedSpendingKey, 
    to: &RecipientAddress, 
    amount: Amount, 
    memo: Option<MemoBytes>, 
    ovk_policy: OvkPolicy, 
    min_confirmations: u32
) -> Result<R, E> where
    E: From<Error<N>>,
    P: Parameters + Clone,
    R: Copy + Debug,
    D: WalletWrite<Error = E, TxRef = R>, 

Creates a transaction paying the specified address from the given account.

Returns the row index of the newly-created transaction in the transactions table within the data database. The caller can read the raw transaction bytes from the raw column in order to broadcast the transaction to the network.

Do not call this multiple times in parallel, or you will generate transactions that double-spend the same notes.

Transaction privacy

ovk_policy specifies the desired policy for which outgoing viewing key should be able to decrypt the outputs of this transaction. This is primarily relevant to wallet recovery from backup; in particular, OvkPolicy::Discard will prevent the recipient’s address, and the contents of memo, from ever being recovered from the block chain. (The total value sent can always be inferred by the sender from the spent notes and received change.)

Regardless of the specified policy, create_spend_to_address saves to, value, and memo in db_data. This can be deleted independently of ovk_policy.

For details on what transaction information is visible to the holder of a full or outgoing viewing key, refer to ZIP 310.

Parameters:

wallet_db: A read/write reference to the wallet database

params: Consensus parameters

prover: The TxProver to use in constructing the shielded transaction.

account: The ZIP32 account identifier associated with the extended spending key that controls the funds to be used in creating this transaction. This procedure will return an error if this does not correctly correspond to extsk.

extsk: The extended spending key that controls the funds that will be spent in the resulting transaction.

amount: The amount to send.

to: The address to which amount will be paid.

memo: A memo to be included in the output to the recipient.

ovk_policy: The policy to use for constructing outgoing viewing keys that can allow the sender to view the resulting notes on the blockchain.

min_confirmations: The minimum number of confirmations that a previously received note must have in the blockchain in order to be considered for being spent. A value of 10 confirmations is recommended.

Examples

use tempfile::NamedTempFile;
use zcash_primitives::{
    consensus::{self, Network},
    constants::testnet::COIN_TYPE,
    transaction::{TxId, components::Amount},
    zip32::AccountId,
};
use zcash_proofs::prover::LocalTxProver;
use zcash_client_backend::{
    keys::sapling,
    data_api::{wallet::create_spend_to_address, error::Error, testing},
    wallet::OvkPolicy,
};

let tx_prover = match LocalTxProver::with_default_location() {
    Some(tx_prover) => tx_prover,
    None => {
        panic!("Cannot locate the Zcash parameters. Please run zcash-fetch-params or fetch-params.sh to download the parameters, and then re-run the tests.");
    }
};

let account = AccountId::from(0);
let extsk = sapling::spending_key(&[0; 32][..], COIN_TYPE, account);
let to = extsk.default_address().1.into();

let mut db_read = testing::MockWalletDb {};

create_spend_to_address(
    &mut db_read,
    &Network::TestNetwork,
    tx_prover,
    account,
    &extsk,
    &to,
    Amount::from_u64(1).unwrap(),
    None,
    OvkPolicy::Sender,
    10
)