Function zcash_client_backend::data_api::wallet::shield_transparent_funds

pub fn shield_transparent_funds<E, N, P, D, R, U>(
    wallet_db: &mut D, 
    params: &P, 
    prover: impl TxProver, 
    sk: &AccountPrivKey, 
    extfvk: &ExtendedFullViewingKey, 
    account: AccountId, 
    memo: &MemoBytes, 
    min_confirmations: u32
) -> Result<D::TxRef, E> where
    E: From<Error<N>>,
    P: Parameters,
    R: Copy + Debug,
    D: WalletWrite<Error = E, TxRef = R> + WalletWriteTransparent<UtxoRef = U>, 

Constructs a transaction that consumes available transparent UTXOs belonging to the specified secret key, and sends them to the default address for the provided Sapling extended full viewing key.

This procedure will not attempt to shield transparent funds if the total amount being shielded is less than the default fee to send the transaction. Fees will be paid only from the transparent UTXOs being consumed.

Parameters:

wallet_db: A read/write reference to the wallet database

params: Consensus parameters

prover: The TxProver to use in constructing the shielded transaction.

sk: The secp256k1 secret key that will be used to detect and spend transparent UTXOs.

extfvk: The extended full viewing key that will be used to produce the Sapling address to which funds will be sent.

account: The ZIP32 account identifier associated with the the extended full viewing key. This procedure will return an error if this does not correctly correspond to extfvk.

memo: A memo to be included in the output to the (internal) recipient. This can be used to take notes about auto-shielding operations internal to the wallet that the wallet can use to improve how it represents those shielding transactions to the user.

min_confirmations: The minimum number of confirmations that a previously received UTXO must have in the blockchain in order to be considered for being spent.