zcash_client_backend::data_api::wallet

Module input_selection

Types related to the process of selecting inputs to be spent given a transaction request.

Structs

GreedyInputSelector

An InputSelector implementation that uses a greedy strategy to select between available notes.

Enums

GreedyInputSelectorError

Errors that can occur as a consequence of greedy input selection.

InputSelectorError

The type of errors that may be produced in input selection.

Traits

InputSelector

A strategy for selecting transaction inputs and proposing transaction outputs.

ShieldingSelectortransparent-inputs

A strategy for selecting transaction inputs and proposing transaction outputs for shielding-only transactions (transactions which spend transparent UTXOs and send all transaction outputs to the wallet’s shielded internal address(es)).