zcash_client_backend

Module fees

Modules

common

sapling

Types related to computation of fees and change related to the Sapling components of a transaction.

standard

Change strategies designed for use with a standard fee.

zip317

Change strategies designed to implement the ZIP 317 fee rules.

Structs

ChangeValue

ChangeValue represents either a proposed change output to a shielded pool (with an optional change memo), or if the “transparent-inputs” feature is enabled, an ephemeral output to the transparent pool.

DustOutputPolicy

A policy describing how a ChangeStrategy should treat potentially dust-valued change outputs (outputs that are likely to be without economic value due to fee rules).

SplitPolicy

A policy that describes how change output should be split into multiple notes for the purpose of note management.

TransactionBalance

The amount of change and fees required to make a transaction’s inputs and outputs balance under a specific fee rule, as computed by a particular ChangeStrategy that is aware of that rule.

Enums

ChangeError

Errors that can occur in computing suggested change and/or fees.

DustAction

An enumeration of actions to take when a transaction would potentially create dust outputs (outputs that are likely to be without economic value due to fee rules).

EphemeralBalance

EphemeralBalance describes the ephemeral input or output value for a transaction. It is used in fee computation for series of transactions that use an ephemeral transparent output in an intermediate step, such as when sending from a shielded pool to a ZIP 320 “TEX” address.

StandardFeeRule

An enumeration of the standard fee rules supported by the wallet backend.

Traits

ChangeStrategy

A trait that represents the ability to compute the suggested change and fees that must be paid by a transaction having a specified set of inputs and outputs.