zcash_client_backend

Module proto

This module contains generated code for handling light client protobuf structs.

Modules

compact_formats

proposal

service

Enums

ProposalDecodingError

Errors that can occur in the process of decoding a Proposal from its protobuf representation.

Constants

PROPOSAL_SER_V1

Constant for the V1 proposal serialization version.