zcash_client_backend::proto

Module proposal

Modules

proposed_input

Nested message and enum types in ProposedInput.

Structs

ChangeValue

A proposed change or ephemeral output. If the transparent value pool is selected, the memo field must be null.

MemoBytes

An object wrapper for memo bytes, to facilitate representing the change_memo == None case.

PaymentOutputPool

A mapping from ZIP 321 payment index to the output pool that has been chosen for that payment, based upon the payment address and the selected inputs to the transaction.

PriorStepChange

A reference to a change or ephemeral output from a prior step of the proposal.

PriorStepOutput

A reference to a payment in a prior step of the proposal. This payment must belong to the wallet.

Proposal

A data structure that describes a series of transactions to be created.

ProposalStep

A data structure that describes the inputs to be consumed and outputs to be produced in a proposed transaction.

ProposedInput

The unique identifier and value for an input to be used in the transaction.

ReceivedOutput

The unique identifier and value for each proposed input that does not require a back-reference to a prior step of the proposal.

TransactionBalance

The proposed change outputs and fee value.

Enums

FeeRule

The fee rule used in constructing a Proposal

ValuePool