Module zcash_client_backend::wallet

Structs representing transaction data scanned from the block chain by a wallet or light client.

Structs

AccountId

A type-safe wrapper for account identifiers.

SpendableNote

Information about a note that is tracked by the wallet that is available for spending, with sufficient information for use in note selection.

WalletShieldedOutput

A subset of an OutputDescription relevant to wallets and light clients.

WalletShieldedSpend

A subset of a SpendDescription relevant to wallets and light clients.

WalletTx

A subset of a Transaction relevant to wallets and light clients.

Enums

OvkPolicy

Describes a policy for which outgoing viewing key should be able to decrypt transaction outputs.