zcash_client_backend
Module wallet

Structs representing transaction data scanned from the block chain by a wallet or light client.

Structs

NoteId

A unique identifier for a shielded transaction output

ReceivedNote

Information about a note that is tracked by the wallet that is available for spending, with sufficient information for use in note selection.

TransparentAddressMetadatatransparent-inputs

Metadata related to the ZIP 32 derivation of a transparent address. This is implicitly scoped to an account.

WalletOutput

An output that was successfully decrypted in the process of wallet scanning.

WalletSpend

A reference to a spent note belonging to the wallet within a transaction.

WalletTransparentOutput

WalletTx

The shielded subset of a Transaction’s data that is relevant to a particular wallet.

Enums

Note

An enumeration of supported shielded note types for use in ReceivedNote

OvkPolicy

Describes a policy for which outgoing viewing key should be able to decrypt transaction outputs.

Recipient

A type that represents the recipient of a transaction output:

Type Aliases

WalletSaplingOutput

A subset of an OutputDescription relevant to wallets and light clients.

WalletSaplingSpend

A type alias for Sapling WalletSpends.