Trait zcash_client_backend::welding_rig::ScanningKey

pub trait ScanningKey {
    type Nf;
    fn try_decryption<P: Parameters, Output: ShieldedOutput<SaplingDomain<P>>>(
        &self, 
        params: &P, 
        height: BlockHeight, 
        output: &Output
    ) -> Option<(Note, PaymentAddress)>;

    fn nf(&self, note: &Note, witness: &IncrementalWitness<Node>) -> Self::Nf;
}

A key that can be used to perform trial decryption and nullifier computation for a Sapling CompactOutput

The purpose of this trait is to enable scan_block and related methods to be used with either incoming viewing keys or full viewing keys, with the data returned from trial decryption being dependent upon the type of key used. In the case that an incoming viewing key is used, only the note and payment address will be returned; in the case of a full viewing key, the nullifier for the note can also be obtained.

Associated Types

type Nf

The type of nullifier extracted when a note is successfully obtained by trial decryption.

Required methods

fn try_decryption<P: Parameters, Output: ShieldedOutput<SaplingDomain<P>>>(
&self,
params: &P,
height: BlockHeight,
output: &Output
) -> Option<(Note, PaymentAddress)>

Attempts to decrypt a Sapling note and payment address from the specified ciphertext using this scanning key.

fn nf(&self, note: &Note, witness: &IncrementalWitness<Node>) -> Self::Nf

Produces the nullifier for the specified note and witness, if possible.

IVK-based implementations of this trait cannot successfully derive nullifiers, in which case Self::Nf should be set to the unit type and this function is a no-op.